# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Habit Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Habit {
  id: number;
  name: string;
  /** Hex color code for the habit */
  color: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface HabitInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
  /** Hex color code */
  color: string;
  description?: string;
}

export interface HabitUpdate {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name?: string;
  color?: string;
  description?: string;
}

export interface HabitLog {
  id: number;
  habitId: number;
  /** Date string YYYY-MM-DD */
  date: string;
  completedAt: string;
}

export interface HabitLogInput {
  /** Date string YYYY-MM-DD */
  date: string;
}

export interface HabitStats {
  habitId: number;
  currentStreak: number;
  longestStreak: number;
  totalCompletions: number;
  /** Completion percentage in last 30 days (0-100) */
  completionRate30d: number;
  /** @nullable */
  lastCompletedDate?: string | null;
}

export interface GridDay {
  /** YYYY-MM-DD */
  date: string;
  completed: boolean;
  /** Intensity level 0-4 (0=none, 4=done, for future multi-count support) */
  level: number;
}

export interface DashboardHabit {
  id: number;
  name: string;
  color: string;
  completedToday: boolean;
  currentStreak: number;
}

export interface Dashboard {
  date: string;
  totalHabits: number;
  completedToday: number;
  longestActiveStreak: number;
  habits: DashboardHabit[];
}